Bug fixes (#85)
* set slotaction swap default to true

* new suggestion required when portfolio changed

* Update version and highlight sell

* change text colours

* fix integer overflow

* fix stale ge open state

---------

+++ src/main/java/com/flippingcopilot/model/FlipV2.java
@@ -51,7 +51,7 @@ public class FlipV2 {
             return 0;
         }
         long gpOut = (spent * amountToClose) / openedQuantity;
-        int sellPrice  = transaction.getAmountSpent() / transaction.getQuantity();
+        int sellPrice  = (int) (transaction.getAmountSpent() / transaction.getQuantity());
         int sellPricePostTax = ProfitCalculator.getPostTaxPrice(transaction.getItemId(), sellPrice);
         long gpIn = amountToClose * sellPricePostTax;
         return gpIn - gpOut;
